Move getJSONFromURL to the transport package
Move it and make it public so it can be reused outside of alertmanager package
