+++ transport/http.go
@@ -1,4 +1,4 @@
-package alertmanager
+package transport
 
 import (
 	"compress/gzip"
@@ -11,10 +11,9 @@ import (
 	log "github.com/Sirupsen/logrus"
 )
 
-// getJSONFromURL is a helper function that takesan URL, request timeout
-// and target structure, it will make a HTTP request and decode JSON response
-// onto the structure provided
-func getJSONFromURL(url string, timeout time.Duration, target interface{}) error {
+// GetJSONFromURL allows to fetch Alertmanager data over HTTP transport and
+// decode it onto provided data structure.
+func GetJSONFromURL(url string, timeout time.Duration, target interface{}) error {
 	log.Infof("GET %s", url)
 
 	c := &http.Client{
