+++ transport/http_test.go
@@ -1,9 +1,10 @@
-package alertmanager
+package transport_test
 
 import (
-	"io/ioutil"
 	"testing"
 	"time"
+
+	"github.com/cloudflare/unsee/transport"
 
 	log "github.com/Sirupsen/logrus"
 	httpmock "gopkg.in/jarcoal/httpmock.v1"
@@ -17,7 +18,7 @@ type mockJSONResponse struct {
 }
 
 func TestGetJSONFromURL(t *testing.T) {
-	log.SetOutput(ioutil.Discard) // disable logging to console
+	log.SetLevel(log.ErrorLevel)
 	httpmock.Activate()
 	defer httpmock.DeactivateAndReset()
 	mockJSON := `{
@@ -29,14 +30,14 @@ func TestGetJSONFromURL(t *testing.T) {
 	httpmock.RegisterResponder("GET", "http://localhost/", httpmock.NewStringResponder(200, mockJSON))
 
 	response := mockJSONResponse{}
-	err := getJSONFromURL("http://localhost/", time.Second, &response)
+	err := transport.GetJSONFromURL("http://localhost/", time.Second, &response)
 	if err != nil {
 		t.Errorf("getJSONFromURL() failed: %s", err.Error())
 	}
 
 	httpmock.RegisterResponder("GET", "http://localhost/404", httpmock.NewStringResponder(404, "Not found"))
 	response = mockJSONResponse{}
-	err = getJSONFromURL("http://localhost/404", time.Second, &response)
+	err = transport.GetJSONFromURL("http://localhost/404", time.Second, &response)
 	if err == nil {
 		t.Errorf("getJSONFromURL() on invalid url didn't return 404, response: %v", response)
 	}
